refactor: 리뷰 반영 (#108)
